+++ src/examples/common/utility.h
@@ -7,12 +7,12 @@
 /**
  * Maps a value from the range [0, 100] to [min, max].
  */
-uint8_t map(uint8_t value, uint8_t min, uint8_t max);
+uint8_t map(uint8_t value, uint32_t min, uint32_t max);
 
 /**
  * Maps a value from the range [min, max] to [0, 100].
  */
-uint8_t pam(uint8_t value, uint8_t min, uint8_t max);
+uint8_t pam(uint32_t value, uint32_t min, uint32_t max);
 
 
 #endif
